Remove allowJs from tsconfig
allowJs breaks ts-loader since typescript cannot parse the flow files.
If we override allowJs from ts-loader then we run into issues with missing
module declarations for our js files. To be consistent we disable allowJs
and use @ts-expect-error on the flow imports which we will remove as we convert
the flow files to typescript.


Former-commit-id: 7a5f0dee5e794c4fecf6325be9e1ca9aee23604f