Fix debouncePromise test (#85)
Some assumptions in this test were broken, likely due to the Node upgrade. Instead of `await Promise.resolve()` to allow the run loop to process async events, I updated the test to use specific signals for different events (such as the beginning of a function call, a promise returning, etc).


Former-commit-id: b1e6c024dc36e606603c368d7dadac70f7b8c0b5