Fix 3D panel circular import and Interaction stories (#146)
The main problem here was a circular import that caused a component to be imported as `undefined`. I found a webpack plugin that can help us detect these—turned on warnings for now, but there are too many to fix currently. Also the plugin doesn't understand that our async imports are a valid way of breaking cycles (https://github.com/aackerman/circular-dependency-plugin/issues/67).
One of the stories had a timing issue — I looked deeper for how to address this, but I think some problems stem from hacky use of `ref` as a way of detecting `onMount` in class components... ideally we would convert everything here to hooks, but that is too big a task for this PR, so just adding a band-aid on this story for now.

I also found that the Layout stories pertain to enabling map height rendering for the Cruise internal map, so are irrelevant here (reverting part of https://github.com/cruise-automation/webviz/commit/6eb6fd9c50b0691e9e5f501644c6957ac9d8d871).


Former-commit-id: 490f6cedf41d65d2e023c57a187a797b2f8ae5fc