Tweaks for flaky stories (#193)
It seems that the OffscreenCanvas doesn't give us a way to know when it has definitely been rendered on screen unless we use `transferToImageBitmap()`, which is a bigger change than I wanted to do for just fixing stories (and before we have perf testing infrastructure). So just add a delay for now.

Former-commit-id: c3a72009d1536369ec7ef187bd5dc1f81b7c9fa5