Convert MessageHistoryDEPRECATED tests to useMessagesByPath tests (#222)
Switch these tests from `enzyme` to `@testing-library/react-hooks`.

Some extra renders are now logged, because MessageHistory wrapped its children in a `React.memo`, but the new API is just a hook that is unable to control if the calling component is memoized or not. Still, I think we are seeing more renders in these simple tests than we should be, and I will file a follow-up ticket to try and improve it.

Former-commit-id: 3c89c44ee79b546ede743dc876da0d36385fd018