Rework Map panel data filtering (#570)
Rather than keeping a Map of _all_ the points in a bag, filter the
blocks (messages from a bag) down based on the current zoom level.

We were previously filtering the displayed points based on the zoom level,
but were also storing all the points as well. This is unecessary.

Former-commit-id: 6711813dad1f8a0e5134d31c646352b2f7304f6b