Subscribe ROS players to /clock, use that for currentTime when available (#529)
* Subscribe ROS players to /clock, use that for currentTime when available

This attempts to subscribe to `/clock` if the topic is available. startTime is set to the first observed /clock message and currentTime is a linear extrapolation from the last observed /clock message. Note that the linear extrapolation will produce strange results if playback is not running at 1x speed. There is a future task once we have rosparam support to check for `use_sim_time` and disable extrapolation if set to true.

Former-commit-id: 089de751cdc2dd059a0d17237957d3cba8a8397d