URDF Viewer panel (#617)
Add a panel to view URDF files that are dropped or double-clicked, or provided on the /robot_description param.

- Drag and drop or double-click a URDF file to view
  - Look up mesh resources relative to the dropped file path
  - Convert TIFF textures to PNG in main process since Chromium doesn't support TIFF
- Load URDF from `/robot_description` param and resolve assets using `env.ROS_PACKAGE_PATH`
- Sliders to change joint values (like joint_state_publisher_gui, although doesn't actually publish the values)
- Dropdown to switch between models from `/robot_description` and dropped file(s?)

This is separate from the 3D panel because it takes advantage of existing `urdf-loader` library which depends on Three.js. We are inclined to work on converting the 3D panel to use Three.js in the future, so the URDF Viewer panel can be a testing ground for initial Three.js integration. Unfortunately this means that visualizing the robot model in world space with other markers, pointclouds, etc. is not yet possible.

Possible future work:
- Buttons to reset / randomize joints (like joint_state_publisher)
- draggable joints from https://github.com/gkjohnson/urdf-loaders/blob/master/javascript/src/URDFDragControls.js

Former-commit-id: 130c3bdb85f66cdab70529afeca303ebfed0c124