Replace react-cytoscapejs with our own component (#644)
Having direct control of cytoscapejs allows us to properly re-layout
the elements when we change them. Previously, changes to the graph
would not result in a layout and the elements would appear stacked on
each other.

This change introduces a new "zoom fit" button to allow a user to reset
the zoom bounds manually.

Former-commit-id: 6351de6e3093b210316320b1449536861765c0e0