Fix unstable useMessagePipeline selectors (#646)
If the selector of useMessagePipeline returns a different object for
each invocation it will cause a state set and re-render. Which will
invoke the selector which would again return a different object and repeat
the cycle.

This change updates useMessagePipeline users to select on a single value
which will mirror that values object identity changes.

Former-commit-id: 15875098030ef3023c935a0c7560978f3fbdb1b1