Rework how players produce warnings and errors. (#670)
Previously, players used sendNotification whenever they encountered a
warning or an error. sendNotification is a one-way out-of-band reporting
mechanism. Once the player produced the error it would persist in the app
(and in the user's app bar) for the remainder of the app lifetime. This
led to a confusing user experience when a player had a transient error
that was resolved (i.e. reconnection after failed connection).

This change removes sendNotification from all players and replaces it
with a _problems_ entry in the player state. The player state is how
players communicate their updates to the outside world. Problems are another
form of player updates. By sending problems via the player state, a
player can add/remove/clear problems as necessary.

This change updates the player presence indicator ui to indicate the presence
of problems in addition to the player presence. The user can see the list
of problems by clicking the presence indicator and and further seeing any
details a problem may expose.

Former-commit-id: 182ed1ade8c9d4ef001bbaa852e542a2bb7dc19a