Make eslint warnings into errors with per-file overrides (#872)
Our remaining warnings come from `@typescript-eslint/no-non-null-assertion` and `@typescript-eslint/no-explicit-any`. This PR changes these to be errors for new files, with an explicit list of files where they are allowed as warnings. This will allow us to gradually burn down the list of overrides as we did for `@typescript-eslint/strict-boolean-expressions`

Former-commit-id: f9ca2bfaed1a9ae825811c41801fda7f84f9d54e