Fix handling of min/max outside of dataset range. (#884)
If a user entered a max value which was smaller than any datapoint value,
chart.js would crash when computing linear scales. This occurs due to internal
chartjs linear tick calculation logic which produces no ticks. As a workaround,
we detect this case and set the min value explicitly which creates a tick.

Fixes #875

Former-commit-id: 421516819b5b1c975a133a26529ce2e4142dbc5a