Account sidebar & Firebase login (#1079)
- Add AuthContext with authentication functions, FirebaseAuthProvider which implements auth by opening `https://foxglove.dev/auth` in an external browser, and FirebaseAppContext which manages the firebase config.
- Add an Account sidebar with UI for the above functionality.
- Handle `foxglove://auth/login-complete` and allow pasting auth token as fallback (protocol handler doesn't work in dev mode on all platforms).
- Convert the app to support a single running instance only, instead of multiple instances, so that the foxglove:// link is handled by the calling app instead of opening a new app.

We have some ideas around testing, but will tackle in separate PRs.

It seems like wrapping the test run in `firebase emulators:exec` is the most well-supported way of having a clean test environment. I don't think it would be easy to spin up separate emulators for each test, but a test can probably use a unique firebase project id so that we are able to run tests in parallel.

Former-commit-id: ce37ee3ecadba90b4cf534d8366a0eddfa3c86ba