Revert "Remove duplicate ForkTsCheckerWebpackPlugin for web build" (#1152)
This reverts commit 023997ad9934a8bac0241f5b14041ef210e24cff [formerly 42d8b4097a885cabbdf80da589afee4eb2f9043b].

This is needed to catch TS errors in files in `web/src/`. By default the ForkTsChecker will look at `web/tsconfig.json`, which is only used for top-level .ts files in the web folder (i.e. `web/webpack.config.ts`) but will not typecheck files under `src`. An example error (with CI green) can be seen in https://github.com/foxglove/studio/pull/1151 ... this is the same error that was fixed in #1109 when I added the ForkTsChecker.

Former-commit-id: 3db0b41c6b3518f5a0f673a8d7a4e3acaad724cb