Add previewTime and setPreviewTime to studio extension api PanelContext (#1200)
The previewTime indicates the time in seconds when a user is hovering
the cursor on the seek bar. The time is within the start/end range of playback.

The setPreviewTime function allows panels to manually set the preview time.

This change adds support for previewTime to the map panel. The map panel
will highlight the most recent gps coordinate leading up to the previewTime.

Fixes #801

Former-commit-id: 178d5c5991fb3718de7b029809db4989646d551c