Add Layouts tab to sidebar (#1246)
**User-facing changes**
Your layouts are now found in the app's sidebar instead of the top bar, with a new look.

**Description**
Replaces the old LayoutMenu with a new LayoutBrowser in the sidebar. The same actions are supported as the layout menu, but with a refreshed UI that will allow us to move towards a world with more interesting layout organization/sharing flows.

Former-commit-id: 393e557c8dcc7f0f955fbaa96a03113d517e8a9a