Update RawMessage panel to allow filtering on any field (#1297)
Previously the RawMessage panel would limit when a user could "pivot"
a field to certain conditions. This change removes those restrictions
and allows a user to select any field within an array of objects as
the filter value.

This change renames the "pivot" tooltip to "filter" as that better conveys
what the action does.

Former-commit-id: 81df7de098f1c8e85f3efbc5c0ba0c99945e7bbe