Fix PointClouds out of bounds read for non-rgb color field (#1559)
**User-Facing Changes**
Fixed a bug where point clouds would fail to render with certain color settings selected.

**Description**
Fixes #681

The PointClouds rendering command used the same shader for rendering a single-valued colorField and a multi-valued (`rgb`) colorField. The shader was hard-coded to always accept `vec3 color`, but only `color.x` was actually used for single-color modes. When a single-color mode was used and the colorField occurred near the end of the point struct, this caused an out-of-bounds read.

To fix the problem, this change separates the shader into 2 versions, one with `vec3 color` for RGB/BGR modes and one with `float color` for single-color modes.

Tested manually with flat, rainbow, and gradient color modes on the original bag from #681, as well as a custom test bag with an `rgb` field.

Former-commit-id: 9b9f950448e2a186f9050fddc742f819d7eac889