Internal: fix formatTime tests (#1697)
formatTime tests for parseTimeStr used the local time zone to create a time
string and then tried to compare that with a parsed value in the America/Detroit
time zone. When a developer is in the America/Detroit time zone this test would
fail since there would be no difference.

This change fixes the tests to use hard-coded time strings and explicit expect
values for sec/nsec. The users local time zone no longer changes the test.

Former-commit-id: b673f23960e5c36ed86d33a4bd12c08e04c34044