Add transition animations to StudioToastProvider (#1842)
**User-Facing Changes**
Not worth mentioning in release notes

**Description**
Closing a toast notification felt slow. While some unnecessary renders are done when closing a toast (particularly because useToasts includes the array of current toasts in its output), the main reason it felt slow is that the toast provider by default includes a transition, but we didn't have any animations for the transition.

This PR imports a simplified version of the transition from the builtin ToastElement.

Former-commit-id: f5e8fbf3a331ac44875b311b8323d833e22678ad