Add mcap package with Message Capture parsing support (#1885)
**User-Facing Changes**
None

**Description**
Adds a workspace package `@foxglove/mcap` with parsing support for the nascent mcap file format. This PR contains no Studio integration, just adding the package since it's begun to stabilize.

The package includes a validate script which can be used to test with local .mcap files:

```sh
$ yarn workspace @foxglove/mcap validate /path/to/file.mcap [--deserialize [--dump]]
Reading gps.mcap
Read 886.10kiB in 89.54ms (9.66MiB/sec)
Record counts:
       1 Chunk
      22 ChannelInfo
   30445 Message
       1 Footer
```

Because the file format and especially the parser is still evolving, we'll keep it in the studio repo for now and later move it into a separate repo.

Former-commit-id: f9761047e6115294dc4f7cdc9477a3cb596046e6