mcap: require messages to be preceded by channel info within the same chunk (#1944)
**User-Facing Changes**
None

**Description**
Message records are required to be preceded by a corresponding channel info within the same chunk. Previously, the parser ignored this requirement and left validation up to the caller. This commit changes the parser to track ChannelInfo records across calls, and changes the Message record type to directly reference the corresponding ChannelInfo (since we are now certain it exists). It also adds logic to ensure the channel info is present within the same chunk as the message, not any earlier chunk.

Former-commit-id: 02cd05173d61a1169b0451a5230d699d15a7b0f9