Disable the CombinedDataProvider (#2006)
**User-Facing Changes**
Loading a second ROS1 bag presents the user with an error indicating the app only supports loading one bag file.

**Description**
When users load multiple ROS1 bag files the more common expectation is that these files would load as one giant bag. The existing behavior - inherited from webviz - would try to load the second bag under a prefix with the idea being a user could compare the bags. For now we comment out this feature to see what the impact is for users. We'll aim to remove this feature and re-introduce it when we re-work data source selection to allow for loading multiple bags as one bag or with a prefix.

Note that ROS2 does not behave this way so this makes bag loading more consistent.

Former-commit-id: 5e6a60fa24ee048aeb94e89c5d35bb879bc1ce3a