WebSocket: Remove leading dot from Protobuf datatype name (#2247)
**User-Facing Changes**
None (feature flagged)

**Description**
This will be needed for the WebSocket player to take advantage of #2246.

I'm not sure if there are any caveats to doing this (like potential name conflicts). My guess is it's fine, since _all_ `fullName`s should have a leading dot, removing it should not cause any name collisions.

Former-commit-id: e435977c55c10cd61b0229c6a07f004653ea032f