Use 1px margin for mosaic tiles next to a splitter. (#2281)
react-mosaic renders tiles adjacent to each other and splitters on top of tiles. A tile without a margin is obstructed by any adjacent splitters. However, having a margin on all tiles results in unsightly gaps and borders. This change adds a margin to tiles only when there would be an adjacent splitter.


Former-commit-id: 9aa30330ce146f07a90fd14aa32ef1c690450074