Add basicDatatypes to UserNodePlayer roslib (#2289)
User nodes are only able to output known datatypes. Prior to this change, known datatypes would be read from the current data source. This limited the types of messages you can output from user nodes to the datatypes in the data source.

This change adds the basic datatypes to the known datatypes for user nodes. This allows a user node author to output any of the basic datatypes even if not present in the data source.

Fixes: #1261
Former-commit-id: b26b5b98752a3e1dcac7fc4875d7fc2456f2f139