Add sample data source (#2362)
**User-Facing Changes**
Part of the new Open dialog experience and not enabled by default.

A new "Explore sample data" button appears on the open dialog. When clicked it loads the sample dataset and layout.

**Description**
A sample data source provides an initialization function to create a player and a layout. When selected, the player is initialized and the layout is loaded to the workspace.

Former-commit-id: 7b091447f23d14743485419c9b1c2f65fc035123