Remove RandomAccessDataProviderDescriptor concept (#2385)
**User-Facing Changes**
None. This is an internal architecture change.

**Description**
Rather than describing a pipeline of data provides abstractly with descriptors, each factory that wants to create a RandomAccessPlayer explicitly creates the DataProviders required for its data source. This allows the factory to create the data provider in a way appropriate for the specific data source rather than requiring all components to adhere to one descriptor specification.
    
This is a precursor to putting more data providers inside workers. Which we need to enable arbitrary file size handling for local ros2 bag files.

Former-commit-id: 6d0e744df4581b5c97f1497d5c32c762bab34977