Add support for recent file (#2388)
**User-Facing Changes**
None. The Open Dialog is feature flagged off by default.

Within the open dialog, the list of recent sources will contain recent files when opening a single file.

**Description**
When a file is opened via the open dialog, we record an entry into recents with the FileSystemFileHandle. This entry can later be clicked to re-open the existing file.
    
On web this feature is restricted to https and may prompt the user to confirm file access before the app is able to read the file again.


Former-commit-id: d418ba8688cf4607da62eb71a1465286daae3147