Add isSampleDataSource to RandomAccessPlayerOptions (#2528)
Rather than detecting a specific url for a demo bag via isDemoBag, this change adds a new constructor option to RandomAccessPlayer for a data source factory to specify if it is for a sample data set. This option updates the metrics collector to tag sample data vs non-sample data. This will support additional sample data sets.

Former-commit-id: a43b4d3b60819b87ce59d10c21ad2d7867351db9