Remove panel name from panel toolbar (#2685)
The panel name shows up in the panel toolbar when a panel is considered wide enough. Even with this logic, the panel name takes up valuable toolbar space when there are other controls in the toolbar.

Seeing the panel name is not important to visualization and diagnostic of the users dataset so we remove it in favor of more space for other UI elements.

Former-commit-id: f77be06da8fa93b89c582bf95ef8254a2cd04bf6