Update PanelExtensionContext#subscribe to handle empty arrays (#2696)
When invoking subscribe with an array of topics, the expected behavior
is to set the current subscriptions to the list of topics. This results in unsubscribing from topics not present in the latest array of topics to subscribe. This was correctly handled for topic arrays of length 1 or more but was incorrectly handled for empty topic arrays.

This change updates the logic to handle empty topic arrays in the same manner.

Fixes: #2679
Former-commit-id: f437859458a148ac32d3e2f9489e9035b801a993