Include default values for protobuf message fields in WebSocket player and MCAP provider (#2735)
**User-Facing Changes**
Default values are now shown in Protobuf messages when using the WebSocket player or MCAP files.
`bytes` now display correctly as Uint8Arrays in the Raw Messages panel when using the WebSocket player.

**Description**
Fixes #2708

By passing `{defaults: true}` to `toObject`, we produce an object that includes all the default values.

When `toObject` was not used, the Raw Messages panel would call `toJSON` (which is defined for protobuf message classes) which would convert bytes to base64. After passing through `toObject`, `toJSON` is no longer defined so Uint8Arrays show up correctly in the panel (and any other app features that relied on `toJSON`).

Former-commit-id: 7fe2badd0bc495cf55470c763572ac78c413eb7b