Update SceneBuilder to check for required PointCloud fields. (#2766)
PointCloud rendering requires field descriptions for x, y, and z fields. This change updates SceneBuilder to check for these fields prior to adding the point cloud to the collector. This allows the UI to display a topic error for the malformed topic better informing the user to errors with their data.

Fixes: #2759
Former-commit-id: bb3f33d0fa261e976d5e2016a529752fe0ac6ed9