Disable tracking of out-of-order messages, which leaked memory (#2799)
**User-Facing Changes**
Fixed a memory leak when the app received many out-of-order messages due to the presence of the `/clock` topic.

**Description**
Out of order messages were being logged to the dev console. The memory debugger showed the console itself was responsible for retaining memory. So we disable logging of out of order messages unless explicitly enabled for debugging purposes.

This is a follow up to https://github.com/foxglove/studio/pull/2185. That fix prevented all the messages from being held in memory by the VM, but they were still held by the browser console when logged.

Former-commit-id: 8b852fe4f5f9755ee9eeb4a86332a8be6e8533c5