Add support to detect existing datatype from node playground (#2831)
When using the `./types` utility file for Message datatypes as output type, node playground detects the original datatype name and sets the node output datatype to the original datatype name. These datatype names matter to other parts of studio which rely on datatype name matching.

This behavior mirrors the existing behavior for the generated "ros" library.

Former-commit-id: 94e6f1b4f40438f57b42375c2649bddab1697a4f