Fix playback bar getting stuck in drag mode (#2912)
**User-Facing Changes**
Fixed an issue where scrubbing on the playback bar only worked as long as the mouse was actually kept on the playback bar, and could get "stuck" if the mouse was released when outside of the playback bar.

**Description**
Fixes https://github.com/foxglove/studio/issues/2900. This behavior was previously working and regressed in https://github.com/foxglove/studio/pull/1805. The mouseup and mousemove events are now tracked on the window instead of the slider element when the mouse is pressed, which improves the UX by allowing a drag to move outside the playback bar and continue working, and allows us to correctly track mouseup even when it happens outside the playback bar (or outside the window). Additionally, a mouseenter/mouseleave pair are used to track the "mouse inside" state, which allows us to hide the tooltip on mouse up, if the mouse was outside the slider.

Former-commit-id: 3abb216500b07983eff9847d73c1e373c4f688e9