Fixes clicking in the upper area of a panel (#2989)
When using a floating toolbar, the hidden toolbar would prevent clicking the upper area of a panel. This regressed in https://github.com/foxglove/studio/pull/2877. This change fixes the bug by setting the width of the toolbar to 100% for toolbars with children and no set width for no children. This lets the panel toolbar container size to the controls rather than across the entire panel.

Fixes: #2987
Former-commit-id: aa1956584e5e493f2a27455d958e28da1c706581