Change package:// url handling to use package: protocol (#2970)
**User-Facing Changes**
When loading a resource via `package://` urls fails, the user is shown an error that contains the failed `package://` url rather than a `x-foxglove-ros-package` url.

**Description**
This change removes the package:// url re-writing in favor of trying to load the package:// url using desktop protocol handlers. This preserves the urls as package:// urls which makes them show up as package:// urls in error messages.

For TIFF files (which are not supported by the Threejs ColladaLoader natively, we convert package:// urls into x-foxglove-converted-tiff:// urls. If there is an error loading the .tiff file we show the unconverted url path to the user to avoid showing the x-foxglove-converted-tiff urls.

Fixes: #2929

Former-commit-id: 313ff5538d565418b3cb440046df866a2dd6ab4e