Prevent rendering CurrentUserProvider until first load (#3131)
The ConsoleApiCookieCurrentUserProvider would prevent rendering children
until the async loading of the user was complete. However, since the currentUser
was actually set in an effect (to allow for unsetting on signout), this would
still result in a render where the currentUser would be undefined followed immediately
by a render where the currentUser is defined.

To avoid this dual rendering, we use a separate state variable to indicate we've
completed the loading. This state variable is set in the same cycle as the currentUser state.

Former-commit-id: 8c36e8c3001c81a89dc2fe9cc277850dd0cc6a0c