Update user-edited node name for next render (#3218)
To ensure the user's cursor remains in the correct position while editing the node name in the node playground panel, we need to use local state for the input control value. This state is updated in the onChange event of the input control so the next render of the control has the newly entered value. This preserves the cursor position while editing in the middle of the node name. If we perform a render using the old value and then render with the new value (i.e. our state updates did not result in the inputTitle changing on next render), the cursor position is lost and placed at the end per react behavior.

Fixes: #3147
Former-commit-id: be962ee3dff82ac7df417b4fa59da568e0298973