Add a validate property to IDataSourceFactory formConfig fields (#3224)
A data source factory can provide a validate function for each field. The validate function allows the factory to intercept new values and return an error if the value for the field is not valid. The open dialog uses this to determine whether to show an error for the field or allow the user to click the "open" button.

Fixes: #3203
Former-commit-id: c8b2533df52f5ba4b4ef98ecf18e0b89c7797e3d