Use bounded iterator ranges with IterablePlayer (#3161)
**User-Facing Changes**
Users of the experimental data platform player are able to use the Studio interface even when their data range does not have any data.

**Description**
Prior to this change the iterable player (what we publicly call the Experimental Player) had a behavior where it would appear to "stall" during initialization and seeking. This stall occurred even on player startup. The player would enter the stateStartPlay method which is responsible for playing a few milliseconds of data to put some messages into the pipeline. To play these messages the player created a message iterator with some start time.

The data platform source would produce a message iterator for this start time and start to stream messages. However, in some cases there might not be any messages to yield from the iterator. In that situation, the message iterator logic would finish the current stream, slide to the next time window and try streaming again. This would continue until the end of the entire requested dataset range was reached with not a single message emitted. Since no message is emitted the player has no opportunity to bail on any of these many stream requests. Even tho it only needed a few milliseconds of data, internally the message iterator went through the entire dataset.

This change updates the IIterableSource interface and player to provide an _end_ time for all messageIterator requests. This allows the player to tell the messageIterator that it need not continue trying to find messages beyond a certain time. The data platform source uses this to avoid making new stream requests after this time in the event no messages are returned.

Fixes: #3158
Fixes: #3063
Former-commit-id: 1f565a95cd1935d7ab931952b2ab40022d3e1ebd