Fix the tick handling for IterablePlayer (#3291)
**User-Facing Changes**
When seeking near the end of the data, the app no longer freezes if there are no messages between the start of the seek and the end of the dataset time.

**Description**
When performing a _tick_ the iterable player would fail to track the correct _next_ start time for an iterator when the _tick_ needs to span two iterators (i.e. the current iterator ends but we still need to read more messages to finish the tick). This change fixes this logic to always set the next start time to the end of the iterator and also clear the lastMessageTime when it is used to setup the new iterator.

Fixes: #3285

Former-commit-id: 79ecaddb92d9741cc09aca6797c837e5f2f5d5ca