Add remote fallback to LayoutManager#getLayout (#3329)
When a user visits a studio url (or deep link) with a specific layoutId, the expectation is that the app will display this layout to the user. The app attempts to load the layout by asking the LayoutManager to getLayout with the layout id. The layout manager consults its local cache and returns the layout with that id.

When working in a team setting, a user may receive a link from a co-worker with a layout id. When the user clicks this link, studio attempts to open the layout id but is unable to because the app has not yet sync'd the user's local cache with the team's remote layouts. This leaves the user with an undesirable message to "Select a layout to get started" rather than showing the user the available layout.

This change alters the behavior of LayoutManager#getLayout to fall-back to fetching the remote layout if a local one is not available. The newly fetched remote layout is inserted into the local cache for future lookups.

Additionally, this change ports the layout debugging controls to MUI to continue our ongoing effort to deprecate fluent. These changes have no user facing impact.

Fixes: #3119
Former-commit-id: 77aa83a78041a4ecd57ae025eda20694370bc16e