Handle google.protobuf.Timestamp/Duration in protobuf messages (#3443)
**User-Facing Changes**
Protobuf messages using `google.protobuf.Timestamp` and `Duration` are now treated as though `seconds` and `nanos` were named `sec` and `nsec`, consistent with other timestamp values in the app.

**Description**
Fixes https://github.com/foxglove/studio/issues/3430

Trade-offs: On one hand, this modifies user data during ingestion and it might be surprising for users to see sec/nsec in the Raw Messages panel when their input data was seconds/nanos. On the other hand, if a user provides a foxglove.X message, that message should behave the same in the app regardless of which framework sent the data.
Former-commit-id: 58d2983ae5501c3c54686bba6a3be31eedfd7c0c