Implement requestBackfill for IterablePlayer (#3522)
The IterablePlayer setSubscriptions implementation would ignore any calls which did not result in actually changing the subscriptions. This behavior resulted in panels which want to subscribe to a topic that is already subscribed from getting the message. The way message pipeline handles this is by invoking requestBackfill after a setSubscriptions call but since requestBackfill was not implemented in IterablePlayer (since setSubscriptions itself changed the state to load when needed), this meant panels did not get messages they needed.

This change fixes this behavior by moving the "backfill" state change logic from setSubscriptions into requestBackfill for the IterablePlayer. In a future world we can remove requestBackfill and have message pipeline understand when subscriptions have not changed since it also has the last player state but for now we adhere to the expected pattern of setSubscriptions followed by requestBackfill.

Fixes: #3481
Former-commit-id: e54f22e5590c9a63660fc39179acfe93503f7027