Limit RawMessage value path filterPath action to primitive values (#3551)
The RawMessage getValueActionForValue function is responsible for computing the path actions (filter message path, plot message path, etc) for a value. When computing value actions for values within an array it would incorrectly try to create a filter path using complex array elements (i.e. {msg=={foo: "bar"}}). This is not supported and for complex elements with bigint values would cause the panel to crash with a "Do not know how to serialize a BigInt" error when the user would hover over their bigint value.

This fixes the behavior of getValueActionForValue to avoid making incorrect filterPaths for complex array elements and no longer crash when hovering over bigint values in the raw message display.

Former-commit-id: 89c0c1416663e16a1a0b7964cf523f9a6220b08c