Prevent concurent state emit in iterable player when setting playback (#3512)
When setting the playback speed for the player the player needs to emit state so downstream components can update with the new playback speed. The iterable player emits the state only when idle, however even when idle the player may be loading blocks and emitting state which would cause concurrent state emits which is invalid in our pipeline.

This change fixes this behavior by triggering a state change back to "idle" if the player gets a speed update when already in idle. The state change will trigger an emit of the player state with the new speed.

Fixes: #3375
Former-commit-id: b839859f5be78ebc0dac5467f77f2901bdcacb6b