Add support for cyclical types (#3563)
**User-Facing Changes**
Trying to view data which has types that reference themselves (valid to do in a proto definition) works without crashing the message path input component.

**Description**
Some functions within message path input used recursion to traverse the RosDatatypes object and build internal structures for message path handling. These functions assumed that no type would have a field who's type was the same as the parent type (i.e. the RosDatatypes structure could be represented with a DAG). Protobuf definitions do allow for a type to have a field which references itself as the type of said field.

Here's an example of a type that references itself:
```protobuf
syntax = "proto3";

message Node {
  string name = 1;
  repeated Node children = 2;
}
```

This change updates the message path functions to keep an array of seen datatypes for each branch of depth-first recursion. The functions check whether the type has already been seen and avoids recursing again on the type if it has. This allows the function to terminate without infinite recursion.


Former-commit-id: f938d5eae51fdd7c84240828ef845e6e735fba9c