Remove duplicate nodeSubscriptions filter in UserNodePlayer (#3737)
**User-Facing Changes**
None

**Description**
UserNodePlayer setSubscriptions creates a set of subscriptions for its child player. These set of subscriptions is a merger of the subscriptions from the setSubscriptions call and the required topics from user node scripts.

The logic it used would limit each topic to one subscription entry. Unfortunately, since we support "partial" and "full" subscriptions on topics this meant that subscription information was lost.

This change removes the topic filtering and passes through all requested subscriptions to the underlying player.


Former-commit-id: 43fbaf1c493cbd7f4d4b2f5c453648a6a429b86e