Display buffering indicator in scrubber (#3803)
**User-Facing Changes**
When experimental players are loading data the playback bar shows a loading animation.

**Description**
Add a `loading` property to the ProgressPlot component. When `loading` is true, the component shows a moving stripe animation in the un-filled ranges to indicate loading is happening.

Add a new `BUFFERING` PlayerPresence for players to indicate they are buffering. A BUFFERING player is already initialized but may not be able to complete playback because it is waiting for data.

Refactor the ProgressPlot to remove canvas. There's no need to involve canvas to display what is typically a handful of rectangles. This is accomplished easier with some divs. The browser scales and renders these divs without having to use autosizing logic.

Fixes: #3777

Co-authored-by: Andrew Carr <andrew@foxglove.dev>
Former-commit-id: 572759420becaa33e535956e0961ebd973914928