Change DataPlatformIterableSource messageIterator to fetch entire range (#3913)
**User-Facing Changes**
Faster loading of preload topics for experimental players.

**Description**
Prior to this change, the DataPlatformIterableSource messageIterator would break up a larger request range into 5 second requests to the stream api. The idea was to avoid making a request for a large amount of data prior to the user needing to play that data since the browser fetch interfaces greedily download the content.

This approach works well for playback but performs poorly for block preloading (i.e. messages for plots). When preloading blocks, we want to load lightwight messages for the entire data source range. By using 5 second requests, a long range results in many 10s or 100s of requests. Since each request has overhead on the network and on the data merging this results in a slow load time for preloaded data.

This change alerts the DataPlatformIterableSource to make a request that matches the start/end range to the message iterator. For preloading this can often simplify down to 1 request for the entire preload range.

To avoid overfetching topics for the playback iterator, the BufferedIterableSource is updated to use smaller ranges for requests. Previously it used one continuous range and relied on the underlying data source to avoid overfetching.

Fixes: #3911
Former-commit-id: 1ef5c37aeeb4e23ce660a69aa2f0fd8c7160ce12