Fix placeholder handling for settings input types (#3997)
**User-Facing Changes**
The type information for panel settings api is more accurate for the placeholder property.

vec2 and vec3 input fields support a new `placeholder` property that is an array of placeholder values

**Description**

Prior to this change the `placeholder` field was present for all settings input types. This change alters the type defunitions to include `placeholder` only in the input types that support it.

This change also adds placeholder array support to vec2 and vec3 input types. It also fixes a bug where vec2 and vec3 would not show the NumericInput elements if values was not specified.

Stories added to test the new behaviors.

Fixes: #3822
Former-commit-id: 84dad3ab6be0017a03077c3685eaed0c0eef2e27