Increase default maxBlockSize for CachingIterableSource (#4125)
The macBlockSize determines how many messages can be in a block before starting a new block. Eviction evicts entire blocks so this is the space freed by evicting a block.

The current default is 5 MB which is an oversight from the original PR that added this default. I meant to set this to something more practical like 50MB since one lidar scan or a few raw images could easily be 5MB. This change updates the default to 50MB

This change also updates the default max total cache size to 1GB in bytes rather than 1e9 bytes.

Former-commit-id: af48cc6e14a3878e1a265a74cc53179fb21bb0b1