Fix manual looping when using the latching player. (#4165)
PR #4045 removed the resumePlay function which was responsible for seeking to the start and playing when the user wanted to continue playing from the end. This change adds that functionality back so when a user wants to "play" and playback is at the end, we seek to the start.

Unfortunately, this logic was being intercepted by logic in RepeatAdapter which would immediately try to pause playback when repeat was off and isPlaying was still true. This change updates the IterablePlayer to indicate it is no longer playing when it has reached the end and updates the RepeatAdapter to remove the isPlaying case since the player should be responsible for that condition.

Fixes: #3899
Former-commit-id: d371c3fce6b710b65ad5b94d96ceb39053c3c95c