Automatically add names to makeStyles calls from tss-react (#4217)
**User-Facing Changes**
None

**Description**

Adds a custom transformer to `ts-loader` that transforms calls to `makeStyles()(...)` into `makeStyles({name: "a useful name"})(...)`. This makes our `tss-react` class names human readable so you can find a pointer back to the code that generated a style.

I also added an explicit name to our Stack component, because it's so commonly used that having a shorter name is better.

Former-commit-id: fa8a9e7008942c7ed5071a649bec71c1a630744f