Add "todo" to list of disallowed words in comments (#4296)
Similar to "fixme", this change adds "todo" to the list of disallowed
words in comments.

Rather than use TODO comments in the code, we create issues for any
follow-up work that needs to happen. Issues are how we review and
prioritize outstanding work. TODO comments are not surfaced and tend to
rot with nothing pushing for their completion.

Some instances of TODO would be better served with a comment describing
the implementation tradeoff so that future readers can learn the
context; this does not need to take the form of a TODO or prescribing
what should be done since circumstances or architecture change.


Former-commit-id: 4898818735f48f9300b96cf0f0a9c2f97d856314