[New3D] Per-point point cloud and laser scan picking (#4295)
This adds a pickInstance() method that is called to render a hitmap of
instance indices for an instanced renderable. This allows individual
points, lines, and primitives to be selected. Renderables can set
`public override pickableInstances = true` and optionally specify a
custom `userData.instancePickingMaterial` that should output
`gl_FragColor` (or equivalent) equal to the 32-bit integer `instanceId`
of the picked instance. If this is not specified, a default shader is
used that will use `gl_InstanceID`. Point clouds and laser scans
currently set a custom shader that uses `gl_VertexID`.

Renderables also need to implement `instanceDetails(instanceId)`
returning a POJO of per-instance details that will be visualized in the
UI.

Fixes #3707

Former-commit-id: cb27922525286c2a85414e7e277b435448ad5c2a