Fix click-to-loop playback behavior (#4379)
Looping behavior is handled by the RepeatAdapter component. This
component has a layoutEffect that runs when active data is changed and
checks if playback needs to restart. To restart playback it calls seek
followed by play.

The iterable player receives the seek request and begins a seek backfill
to load messages for the seek time. While it does this, if the layout
effect in the adapter runs again (possibly due to a previous state emit
finishing), then the logic to seek to the start is run again. The
iterable player is again told to seek to the start which interrupts the
previous unfinished seek and the cycle repeats. This leads to the player
never actually seeking to the start.

This change updates the IterablePlayer to check if the requested seek
time matches the current targetted seek time and avoid entering another
seek state and letting the current one finish.

Fixes: #4352

Former-commit-id: 71c2b80dc126ef24aaf3eb7562c491c0f707a722