Support loading files which contain messages spanning decades (#4398)
The BufferedIterableSource is responsible for wrapping an
IIterableSource and performing "look ahead" fetching in the "background"
with the purpose of fetching upcoming data before the player actually
needs to read the message. The background loading logic would request a
message iterator for a fixed amount of time (the readahead window of 5
seconds) from the underlying source. As the player consumed messages
from the message iterator of the BufferedIterableSource instance the
background process would move forward and start new message iterators as
needed to fill the readahead buffer. It knows it can stop buffering new
data when it reads a message that is beyond the required time of the
readahead window.

When loading a file that contains a massive gap in time (decades), this
logic of reading a fixed amount of time fall apart. The
BufferedIterableSource effectively goes into a tight loop. It wants to
read 5 seconds into the buffer. To do this it keeps trying to read 5
second increments from the underlying data source but since the next
message is decades away this creates a tight loop.

I originally wrote the BufferedIterableSource to work in fixed
increments on the underlying source because the DataPlatformIterable
source would otherwise try to fetch too much data over the network too
soon. I now believe this is a problem specific to the
DataPlatformIterableSource and behavior that should be internalized
within that source rather than impacting how buffering is done for all
sources.

This change updates the BufferedIterableSource to create a single
messageIterator on the underlying source and read from that as necessary
to keep the readahead buffer filled to the appropriate time. This avoids
the tight loop because the messageIterator is responsible for producing
the next message from the source regardless of how far from the previous
message it might be.

This means the sliding window reading logic needs to move into the
DataPlatformIterableSource (because we want to avoid making an
open-ended _fetch_ request since that will download the entire range
even if we don't play the range). Unfortunately, moving the sliding
window logic means the same "looping" behavior now happens in
DataPlatformIterableSource. When you open a stream that has gaps, the
source will make many 5 second requests only to learn there are no
messages and move on to the next 5 second section. The fix for this is
to use the _coverage_ information and skip querying sections of time
which have no data.

And finally, when preloading data in the BlockLoader the app wants to
load certain topics for the entire range to display in plot panel or for
3d panel tf preloading. When reading these messages from
DataPlatformIterableSource, we want the app to avoid making many
separate 5 second requests for the data since this is much slower than 1
request for the entire range. To support this (while also avoiding
making 1 request for playback use), we introduce a `consumptionType`
field to the `messageIterator` args. This field can be set to `full` to
indicate the caller plans to read all the data from the iterator in one
continuous pass or `partial` to indicate the iterator will be read
incrementally during playback. The source can use this argument to
change how it wants to load data.

Fixes: #4396
Former-commit-id: 3c55fafcfffc36bae95d642fa08fe9f9e0eb1c9f