fixes to old3d->new3d upgrade flow (#4532)
**User-Facing Changes**
Fixed issues with upgrading to new 3D panel.

**Description**
Fixes #4530 by converting radians to degrees when migrating camera
settings.
Resolves #4531.
Also changes the localStorage key used for the setting so that the
banner will re-open for anyone who has clicked it in the past. This will
help with #4531 to encourage people to click the banner on other old3d
panels even if they've already converted one.

Former-commit-id: d04d3c97362cb3c225f3b5babbf9bb1423f30157