[3D] fix auto-selection of rgbByteOrder (#4685)
**User-Facing Changes**
[3D] Fixed a bug with RGB point clouds where the automatically-selected
settings would sometimes change after the point cloud was initially
displayed.

**Description**
Fixes #4680

We were using the `autoSelectColorField` function to choose the
colorMode and rgbByteOrder, but the rgbByteOrder wasn't actually being
saved in the panel config. So when the renderable got updated, it would
re-read the saved config value, which was undefined, and fall back to a
different value.

Former-commit-id: a2f9fbd991909c3439d64da27dba067e838c76af