Fix erroneous duplicate channel warning in websocket player (#4837)
The existing logic adds a player problem if we see an advertised channel
twice. We really only need to show this if two different channels are
advertised on the same topic. The change here is to compare the new and
previous channels and only warn if they have different properties.


Former-commit-id: 0dba3d88173d5d8f5fc1f2c66d5e8fbf1c9e77b5