Add registerMessageConverter to the extension api (#4507)
**User-Facing Changes**
A one-liner for the release notes:
> Add a new kind of extension for registering message converters to
convert from one schema to another.

The user-facing changes are described in detail within this website PR:
https://github.com/foxglove/website/pull/850

**Description**
Adds a new function `registerMessageConverter` to the
`ExtensionContext`. This function allows a for creating extensions to
tell Studio how to convert one schema to another. Such extensions are
useful for adapting studio visualizations to display custom messages.
For example converting a custom GPS message into `foxglove.LocationFix`
so it can be visualized in the Map panel.

```typescript
registerMessageConverter({
  fromSchemaName: "my.CustomGpsMessage",
  toSchemaName: "foxglove.LocationFix",
  converter: (message) => { ... }
});
```

Registering a message converter alters what is possible for panels to
subscribe to. Assuming the above message converter is registered a few
new things are available to extension panels:

A new `additionalSchemaNames` field is available for any topic having
the `my.CustomGpsMessage` schema. So if we have a topic `some/topic`
that has the type `my.CustomGpsMessage`, we'd have an entry for
`foxglove.LocationFix` in the `additionalSchemaNames` array for the
topic.

```typescript
{
  name: "some/topic",
  schemaName: "my.CustomGpsMessage",
  additionalSchemaNames: ["foxglove.LocationFix"],
}
```

A panel can then subscribe to the topic and request the _converted_
messages by using the `convertTo` subscriber option.

```
context.subscribe([{ topic: "some/topic", convertTo: "foxglove.LocationFix" }]);
```

And finally, any `MessageEvent` the panel receives as a result of a
convertTo subscription, either in `currentFrame` or `allFrames`, would
have a the converted message in the `message` field and a new field
`originalMessageEvent` which contains the original un-converted message
event.

The logic for these additions is implemented entirely within the
`PanelExtensionAdapter` `buildRenderState` function and available only
to panels built using the extension api.

Fixes: #4730
Former-commit-id: 6083f0a442106e1a656e50dff1f6f2b3e1118a73