Use batch requests to fetch messages from WorkerIterableSourceWorker (#4891)
The data source logic runs in a web worker and exposed via the
WorkerIterableSource class. When using the messageIterator method of the
IterableSource this results in a worker RPC call to fetch each next
message. With high rate sources (multi khz) this creates a lot of
overhead to read each message and drastically slows down the buffering
performance.

This change improves the performance substantially by introducing a new
`nextBatch` method to the cursor interface. The `nextBatch` method
allows fetching a batch of messages rather than a single next message.
The `nextBatch` method receives a hint on how many messages to batch.
Since the fastest Studio renders is 60 hz, this means we typically work
in batches of 16 millisecond message ranges. This change updates the
`messageIterator` call for the `WorkerIterableSource` to request batches
of 16 milliseconds from the worker. This reduces the number of RPC calls
from potentially several thousand to ~60 to fetch 1 second of data.

Fixes: #4857
Former-commit-id: 0be748e6b0a615323ba6ba23cfac8f25b6006026