Remove deprecated `datatype` field in MapPanel (#4920)
I cannot collapse this logic down to only subscriptions because we need
the list of eligible topics for showing the settings with all topics. I
did remove the datatype field by updating the expected type for building
the settings tree. I have to omit the field because it is required in
the @foxglove/studio type definitions.

Fixes: #4879

Former-commit-id: 0d8086e435be9ed97a4ecaac90ca959c750e7351