Fix `selectRecent` memory leak keeping old players in memory (#4924)
**User-Facing Changes**
- fix frequent memory leak where `selectRecent` would keep the players
in memory after selecting a new source

**Description**

Because the callback is exposed to the larger PlayerManager closure
context it captures the player in it. This function is used in memoized
state in `Start.tsx`, but the `OpenDialog` that contains it isn't
cleaned up by react on unmount for an unknown reason. `selectSource`
does not seem to have the same problem with its usage. As a workaround,
moving the definition of the `selectRecent` callback to an outside
function will prevent it from keeping the player in its closure context
this way.

The `stateNode` holding onto the detached element is the mui-modal in
`OpenDialog` which is unmounted, but it's still stored in react internal
state functions (ie: `stateNode` and `lastEffect`), which makes me think
this might be a react issue. When we update react we should probably
check to see whether this is still the case because apparently cleanup
in react 18 is more aggressive.

Retainer stack:
<img width="892" alt="image"
src="https://user-images.githubusercontent.com/10187776/205985251-d815fccb-4ab8-45dd-8192-5a6131b6ceb5.png">


To reproduce prior to fix:
(with memory panel open in chrome devtools)
1. use studio default layout
2. Refresh the page and select a local file source
3. Wait for buffering to finish
4. Go to data source tab and select the same local source
5. Take a snapshot in the memory panel
6. See two iterable players in memory, the old IterablePlayer will be
the one that has a greater distance, click to inspect its retainers
7. It's primary retainers should resemble the screenshot above


<!-- link relevant github issues -->
<!-- add `docs` label if this PR requires documentation updates -->

Former-commit-id: e90fd930d3d58b3bbc99ef0a6ef1cf5db085477a