Delete useCleanup hook (#4991)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
 - none

**Description**
Remove useCleanup and usage hook because of issues: It is too easy to
miss-use because it requires a stable function passed to it but does
nothing to ensure the caller passes a stable function. On top of this it
will retain the closure context of the initial callback function until
the component unmounts. Potentially causing temporary memory leaks


<!-- link relevant GitHub issues -->
Fixes #4895
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

Former-commit-id: 015aa734769f37e24fb8cef700b057fc0c22c318