Fix IterablePlayer and UserNodePlayer bugs with seeking (#5047)
**User-Facing Changes**
- Fixed an issue where the app would sometimes not respect the `time`
URL parameter.

**Description**
- Fixes IterablePlayer to support seek events that happen during
initialization (such as the seek from `useSyncTimeFromUrl`), and to not
use `{sec:0,nsec:0}` as a placeholder time before initialization is
completed.
- Deletes a line from UserNodePlayer which was sometimes triggering a
seek with the incorrect timestamp (if another seek had been requested
after `lastPlayerStateActiveData` was captured). The code being removed
was ostensibly supposed to update the user script output when the script
was edited, but that's not working
(https://github.com/foxglove/studio/issues/4989 / FG-1019) so it's
simpler to delete this line. The way script updates are handled will
need to be re-worked to solve
https://github.com/foxglove/studio/issues/4989 and FG-1071.
Former-commit-id: 17279103c53c139ab18666510496b87039c8c4a3