Saving User scripts updates subscribed panels (#5071)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- Saving a user script will update subscribed panels with new output
messages

**Description**

Added state to remember last of each topic's message in the
userNodePlayer and every time `setUserNodes` is called as an update to
an existing node, that node will be added to a set of nodes that need to
have their outputs refreshed. In the next `_onPlayerState` call this
will read in the previous messages for each input topic and add them to
the messages to be processed in `_getMessages` and then remove them from
the list of nodes to update. `IterablePlayer` triggers this during
`idle` state. This avoids an additional `_emitState` call and issues
from my previous solutions with race conditions where previous
playerStates were being overridden and old sourceCode was being shown.


<!-- link relevant GitHub issues -->
Fixes #4989
FG-1019
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

Former-commit-id: 01667857301a91a59fb2529b8a15f6ad9f361a5c