Parameter panel: Add button to submit parameter changes (#5116)
**User-Facing Changes**
- Parameter panel: Add button to submit parameter changes

**Description**
In #5095 I had several issues with the parameter panel:
- When the server rejects parameter updates, the parameter panel should
reflect that. However, the last user-set value is shown
- There is no dedicated submit button and new parameter values are send
while typing, often leading to invalid or incomplete parameter values
being send to the server

This PR adds a dedicated submit button to send paramter updates only
when the user is presses it. Additionally, the parameter row is
re-rendered if the value changes.


![image](https://user-images.githubusercontent.com/9250155/211087861-52a1b5bd-5d35-47ae-a312-78dbd7ac0f56.png)

Former-commit-id: 65cf4f1e076388f99aab0a1717a9e5f68495a483