Support inline actions on settings nodes (#5255)
**User-Facing Changes**
Make it possible for panel authors to request that node actions be
rendered inline instead of in the context menu.

**Description**
Expands the `SettingsTreeNodeActionItem` to include an optional
`display` property, which can be set to `inline` or `menu`. If `inline`
it will be rendered inline alongside the node label instead of in the
context menu. This should make frequently used actions more efficient
and discoverable.


Former-commit-id: 8a9aab7ed8dfa4a30920c5cdffe3dfdff4ea1db0