Fix `web-integration` check to run for forks (#5264)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
 - n/a

**Description**

Remove github actions cache to allow for forks to use this web
integration test. I believe that this is the issue since this is the
only part of the job that depends on a GITHUB_REF. There is open issue
about this on actions/cache. When that is closed we should be able to
add this back.
https://github.com/actions/cache/issues/319

<!-- link relevant GitHub issues -->
FG-1472
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->
Former-commit-id: 0eca15af2fcc97df9f18b6800a5b6a11c5476664