Adding support for YUYV(YUV2) aside from YUV422(UYVY) along with UYVY alias (#5292)
This commit is to add support for YUYV(YUV2) encoding aside from YUV422
which is encoded in the UYVY format. Our cameras take pictures in the
YUYV format and this commit is to make it directly compatible with our
images without having to do conversions with the data. The name might
need to be changed in the future as YUV422 is supposed to be YUYV, not
UYUV, but we did not want to mess up other people's programs through
this change.

Former-commit-id: 920a49d9102254c5d76a71bf07b9f77861d142f9