Add an error boundary to built-in panels (#5432)
When a built-in panel has an error in the react tree this error would
bubble up, log to the console, but never result in a panel error
boundary to show. This change updates the built-in panels with a new
error boundary which relays this error back to the main app react tree
where it can be handled by the panel error boundary we already have.

This change does not address 3rd-party panels which cause unhandled
error. The solution for those is to move extension panels into an
iframe.

Related issue which surfaced the gap in error handling:
https://github.com/foxglove/studio/issues/5426
https://github.com/foxglove/studio/pull/5427
Former-commit-id: 54686bbb0980397b3d32f4903ba7a0cc0d57073c