Always emit urlState in FoxgloveWebsocketPlayer (#5454)
FoxgloveWebsocketPlayer was not emitting urlState when it did not have a
connection. This prevented downstream logic from updating the URL with
the latest player information. This chagne updates the player to always
emit urlState which causes the app to update the URL bar immediately
once the data source is selected.

Former-commit-id: 4e92ae1dd7c887e1a8c3f426034aa7aa4c3a3d54