Exit the producer loop when aborting during stamp wait (#5466)
For some data sources, large gaps in data produce `stamp` results in
their message iterators. These stamp results indicate where the
underlying data source has moved its cursor. The BufferedIterableSource
compares such stamp results to the current playhead location to
determine whether it needs to buffer more data from the data source. The
logic for this comparison sits in a loop until the playhead indicates it
is time to read more data. This loop did not properly handle abort
signals which resulted in the "producer" thread of the
BufferedIterableSource never exiting.

This stall in the BufferedIterableSource was obfuscated by logic in
IterablePlayer which did not wait on the return value of the message
iterator from the BufferedIterableSource instance. This resulted in the
stall occurring on transitions through the next seek, backfill, and play
states of the player rather than during the initial transition to
backfill. This change updates the IterablePlayer logic so it waits on
the existing message iterator to exit before continuing to process the
new state. This makes it more obvious if the player gets stuck trying to
transition to the new state. It also removes an instance of promise
handling that did not await the promise value which contributed to
making it harder to track down the issue.

Fixes: #5380
Former-commit-id: 24c8fc3b9b507c8fa5345efd78121e5bd2aeb63d