Fix some image panel state issues (#5478)
**User-Facing Changes**
Fixed an issue where the Image panel would sometimes continue displaying
annotations after they were disabled in the panel's settings.

**Description**
Addresses two issues:
1. As reported in #5371, the panel would sometimes display annotations
after they were disabled. This happened because of a race condition
around `context.subscribe()` (quite similar to #5473, but for
subscribing rather than publishing), which causes the panel to
occasionally receive a message in `currentFrame` after it has already
unsubscribed. To work around this I added a check that the annotation's
topic is in the list of `annotationTopics` the panel expects to receive.
Also filed #5479 to follow up on the issue.

2. I also found that the panel would sometimes display an image and the
"waiting for images" message at the same time. This was due to some
logic that would display the `lastImageMessageRef.current` if no image
is available. Reading `lastImageMessageRef.current` during render was
wrong, since the modification of the ref does not always trigger the
panel to re-render. Since the `useImagePanelMessages` hook now has
responsibility for keeping the latest image message around, I removed
uses of `lastImageMessageRef` other than for the "download image"
feature.

Fixes https://github.com/foxglove/studio/issues/5371
Fixes FG-1924

Former-commit-id: 65f753e9cfd4826f2e0062ab7c2ebd8e16cc5826