Fix subscribing to multiple converters in single panel (#5510)
`renderState` is responsible for figuring out the `currentFrame` and
`allFrames` fields for extension panels. Part of this logic tracks which
subscriptions need to leverage message converters and runs the
converters. It would track this information by storing a map from the
topic name to the target schema.

Unfortunately, this architecture meant that the panel could only
subscribe to one converter per topic. If the user installed an extension
which registered two converters, each for the same _from_ schema but
different _to_ schemas, the panel could not subscribe to both _to_
schemas as one converter subscription would override the other.

This change fixes this behavior by tracking all of the subscribed
converters for a specific input topic and schema. Now if a panel
subscribes to the same topic with different convertTo parameters, both
of these subscriptions are honored and the panel will receive both
converted messages.

Fixes: #5487
Fixes: FG-2321
Former-commit-id: a9da52da870e21f1083cfaa9983c29a7558dee1c