fix PanelLayout re-mounting panels unnecessarily (#5567)
The ordering of react effects and render calls was causing
`panelComponentCache.current.clear();` to happen after the mosaic tile
had been rendered, meaning on the next render (e.g. when resizing a
panel) the panel would be re-mounted (since the
`React.lazy(panelInfo.module)` was called again). This fix switches to a
`useMemo` for building the map of `React.lazy` components from the panel
catalog.

Relates to #5551 / FG-2498

Former-commit-id: 8a450382d3f88bfd9092d325e1a7f23822b189e4