Fix crash with "Do not know how to serialize a BigInt" (#5578)
**User-Facing Changes**
Fixed a crash in the Raw Messages panel in certain cases involving
64-bit integers nested inside objects nested inside arrays.

**Description**
Check that the nested value is a primitive before suggesting it as a
filter. This avoids a crash when there was a field named e.g. `id` which
contained a nested object that had a bigint somewhere inside it.

Test data:

[test_int64_0.mcap.zip](https://github.com/foxglove/studio/files/11006052/test_int64_0.mcap.zip)

Fixes FG-2528

Former-commit-id: 2068b4e503a6ba27216369b7fc731abad25dce31