Fix 3d panel image scene extension (#5555)
**User-Facing Changes**
- Fix 3d panel bug where visualizing an image would not work if there
were multiple images for the same camera info.
- Fix 3d panel bug where changing the camera info of an image to a
camera info on a different frame would not work.
- Fix 3d panel bug where receiving a camera info message before an image
message would not display the image

**Description**

This change fixes several 3d panel bugs with the Images scene extension.

1. This change fixes a bug where the 3d panel would fail to subscribe to
a camera info topic if there were multiple image topics related to the
same camera info topic and the last image topic (alphabetically) was not
visible but an earlier image topic was visible. This change corrects
logic in `cameraInfoShouldSubscribe` to subscribe if any image topic
needs the desired info topic.
2. This change fixes a bug where the 3d panel would use the `frame_id`
from the `image` message rather than the `camera info` message. This bug
resulted in the image being displayed with the correct camera
information and distortion geometry from the selected camera info but in
the wrong coordinate frame.
3. This change fixes a bug where the 3d panel would fail to display an
image message when received after a camera info message. The bug
happened because the `handleImage` call did not have logic to calculate
a camera model from the latest camera info message. This logic only
happened in `handleCameraInfo`. This change updates the
`handleCameraInfo` logic to save the last camera info message on each
topic so `handleImage` can use it to create a camera model when the
image message is processed.

Fixes: FG-2500
Fixes: FG-2501
Former-commit-id: 062acbe2e9735ec2c219d1ddc36785da85723b96