Don't crash panel settings for unknown panels (#5660)
**User-Facing Changes**
Fixed a crash when viewing panel settings for an unknown panel type.

**Description**
"Unknown panels" can occur when old layouts are loaded, or when
extensions are uninstalled, so the app no longer has a panel component
for an ID that's present in the layout. Since we sometimes do this
intentionally (e.g. #5634) it seems better to not throw an error in this
case.

Former-commit-id: 7cfe12935364d3148df638ab06db599f945d5388