[3D Panel] Fix backfilled messages from disappearing when changing subscription (#5712)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- Fix objects in 3D disappearing when playing after changing topic
visibility

**Description**
Early return when the currentTime hasn't changed caused the seek to not
be handled and set back to `false` in local state. This caused the
behavior where the next time the current time change it would handle the
seek, which clears renderables in the renderer. This would be after the
backfill messages had been processed causing the previously backfilled
messages to be cleared from the renderer.

<!-- link relevant GitHub issues -->
FG-2779
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

Former-commit-id: 1461ddf391a23b8b4354b73c9f7ac2e791baff26