Deserialize protobuf Timestamp/Duration seconds as number rather than bigint (#5757)
**User-Facing Changes**
Fixed an issue where `time` and `duration` types in Protobuf data were
deserialized differently from other encodings, which could lead to
errors in the app.

**Description**
This regressed in #5727, where we changed all int64 to be deserialized
as bigint, without realizing that this also included Timestamp and
Duration built-in types. Studio depends on those being `{sec: number,
nsec: number}`, but after #5727 they were `{sec: bigint, nsec: number}`.

This change adds a postprocessing step to convert the `sec` field to a
number, and throw an error if the value is too large to represent
without rounding. This replaces some old logic which simply renamed the
fields.

Fixes FG-2916
Former-commit-id: 660985ee7fb002530647a8dfe763d89964d9cb11