Fix multi-topic playback for unindexed local mcap files (#5787)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- Fix multi-topic playback for unindexed local mcap files

**Description**

Unindexed MCAP files use the `McapStreamingIterableSource`. The message
iterator returned from this source yields message values by channel
first, then time-sorted within the channel. This cause messages to be
iterated through by the player to also be in that order. So it would
only read messages into the current frame from the first subscribed
topic and no other ones because they were after all of the messages in
that first topic



<!-- link relevant GitHub issues -->
FG-2863
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

Former-commit-id: d93ba64783badba7e4b6b8fcd133a111d98b6430