Update picker to work in image mode (#5860)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- none (experimental)

**Description**
Now in ImageMode the picker does not use a smaller render target.
Previously the `setViewOffset` was overriding the manually set
projection matrix in the camera. The setViewOffset was used to render
only a subset of the view from the camera, which improves performance
when picking. From observation it looks like this makes the
`clickHandler` in ImageMode take two times longer when picking (up from
~50ms to ~100ms) for the same size windows. I think this difference will
be negligible for users when picking. If it does become a problem we can
revisit the manual setting of the camera matrix in image mode.


<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
FG-3096
<!-- add relevant metric tracking for experimental / new features -->

Former-commit-id: ed1a55eccd4193538d91b070cadeeff6e8f82f13