Rework the TimeBasedChart downsample code organization (#5869)
Rather than having two layers of throttle and debounce, this change
moves uses one layer of throttle and a Downsampler class to manage the
state for the downsample. useEffects update the Downsampler instance
with the latest state information and the debounced callback triggers a
downsample which updates the state variable with the downsampled
dataset.

Former-commit-id: 6767f2ff911d5df800aa98e664933430087a23e9