Fix urdf collada mesh color being overwritten (#5873)
**User-Facing Changes**
- Fix Collada meshes being displayed with wrong color

**Description**
Collada mesh materials were ignored if the URDF visual had a an optional
material specified. In Rviz, the optional material is only used if the
mesh itself does not contain texture information. This PR makes sure
that for collada files, embedded mesh materials are always used. To my
knowledge, Collada is the only mesh file type used for URDFs which
contains texture information.

Former-commit-id: 138d7c06fcf55bbed2bb631fc58b87bd9a5a4c7b