Fix Display Frame Error on refresh (#5964)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
 - [3D Panel] Fix Display Frame Error on refresh sometimes


**Description**

The issue was that the Renderer was setting `this.renderFrameId` using
`getDefaultFrame` and it was going to render fine, but the settings
nodes weren't getting regenerated so it wasn't clearing the error
anymore. So any change to the settings (like moving the camera) would
remove the error.

I updated frame fallback conditions to hopefully be a bit clearer. The
display frame settings will only ever display the value of followTf if
it exists and if not it will show the followFrameId. The setting of the
followFrameId is handled in CameraStateSettings so that it can be
updated with the settings node and on transform tree changes. I've
removed `renderFrameId` from the renderer to clear up


<!-- link relevant GitHub issues -->
FG-2985
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

---------

Co-authored-by: Jacob Bandes-Storch <jacob@foxglove.dev>
Former-commit-id: 709459abbee8be35506225f1d8e7107c4036c020