Fix point hover logic on map panel (#6121)
**User-Facing Changes**
Fix logic to detect hovered points on map panel.

**Description**
Fix logic to detect hovered points on map panel. When we have a preview
time the existing logic looks for the most recent point _before_ the
hover time, which causes it to pick the point before the actual hovered
point. The fix is to look for the most precent point _before or at_ the
current hover time which includes the currently hovered point.


Former-commit-id: 54a1534574d97a5bcb60083d7065b507af0daa1d