Fix built-in datatypes overwriting player datatypes (#6193)
**User-Facing Changes**

- Fix built-in datatypes overwriting player datatypes

**Description**

From
https://github.com/foxglove/studio/issues/6190#issuecomment-1576719837

> The reason for the shift is, because Studio uses the ROS1
std_msgs/Header definition instead of the ROS2 definition which should
actually be used. The ROS1 definition contains an extra uint32 seq field
which causes the shift.
>
> The ROS1 definitions are added
[here](https://github.com/foxglove/studio/blob/181414339a42044d30817353052be3dda1fbb6d9 [formerly 54a1534574d97a5bcb60083d7065b507af0daa1d]/packages/studio-base/src/util/basicDatatypes.ts#L18-L26)
and communicated to the publish panel via the UserNodePlayer.

This patch fixes that player datatypes are being overwritten by builtin
datatypes with the same name (e.g. `std_msgs/Header` in this specific
case).

Fixes #6190
Former-commit-id: 9a009234ed9a0af8a117f7abdc4c5edeb9266e01