Update time/duration in protobufDefinitionsToDatatypes to matching deserialized object (#6249)
**User-Facing Changes**
Fixed the `google.protobuf.Duration` and `google.protobuf.Timestamp`
generated type definitions in User Scripts so they match the way
timestamps are represented in the rest of the app (`{sec, nsec}` instead
of `{seconds, nanos}`).

**Description**
We modify the deserialization logic for Duration and Timestamp so we get
`{sec, nsec}`, the same format as the rest of Studio expects. However,
the generated types for User Scripts still showed the names as
`{seconds, nanos}`.

Former-commit-id: 33f09ba9e13323bd59a4f7b27daab28b749eb188