[Map Panel] Display covariance as ellipse (#6281)
Currently, the map panel displays the accuracy of `NavSatFix` messages
as a circle around the point, based on the max variance seen. There's
enough information in the covariance matrix for us to display these as
ellipses. We handle two cases of this, depending on the [covariance
type](https://github.com/foxglove/studio/blob/94b063b7b3a3b4237823028cfdb6671285676988 [formerly 50cf7b623ba307c5e5f5ccc77972ce7c5aeaecc4]/packages/studio-base/src/panels/Map/types.ts#L17-L22).

This adds
[leaflet-ellipse](https://www.npmjs.com/package/leaflet-ellipse) (Apache
2.0) as a dependency to render the ellipses.

The `getAccuracy` function has been moved to a separate file for
testing, and extended in the following ways:

1. For the *diagonal known* type, we return both covariances as the
radii for the ellipse, rather than the max of the two
2. For the *known* type, we use the two most significant eigenvalues
from the matrix, and the eigenvector corresponding to the most
significant value to determine the orientation.

Fixes: FG-3873
Former-commit-id: dfd3261c643309733b245e6f7723765d59afde7e