Delete LayoutManager (#6364)
This change simplifies the CurrentLayoutProvider to manage the layout
state without dealing with saving and loading. This responsibility is
pushed to downstream components. The CurrentLayoutLocalStorageSyncAdapter is updated
to save/restore the current layout to/from localStorage.

The PlayerManager is also simplified to avoid load/save behavior and
instead this is pushed downstream into the CurrentLayoutLocalStorageSyncAdapter
which is updated to load any sample layout into the
CurrentLayoutProvider.

Resolves: FG-3348
Resolves: FG-3031
Resolves: FG-4151
Former-commit-id: 86a323ef29b7156d7d73ce0973bc7fc468b034d0