Avoid re-processing messages on global variable changes (#6391)
Earlier this year @jtbandes filed an internal bug report that
user-scripts do not re-run when global variables are changed which was
the inherited behavior from webviz that we've had since Studio was
created. This resulted in a bugfix PR
https://github.com/foxglove/studio/pull/5173 which re-ran the scripts
when variables changed.

However, this re-running of scripts behavior broke a message pipeline
invariant. The invariant is that we do not re-emit already emitted
messages into the message pipeline. This invariant went un-noticed but
did trigger a player problem from the MessageOrderTracker when you had a
user-script and made changes to global variables (i.e. selecting an
object in the 3d panel). This would cause the user-scripts to run and
re-emit already emitted messages and appear to panels like data went
back in time.

We put together a workaround to silence this player problem
https://github.com/foxglove/studio/pull/6164 but the invariant behavior
remained. After an internal discussion we decided to revert the original
PR which introduced the invariant behavior and wait for examples of
real-world workflows around the variable behavior before considering it
a "bug" that scripts do not emit when variables change. The current
script behavior is to emit when input messages change (or the script is
edited and saved).

This change reverts #5173 and the `messagesRecomputed` changes from
#6164.

Fixes: FG-4059
Fixes: FG-3749
Former-commit-id: 75d4980a372d4cbe5c3ca709858e76904d56d578