Remove SignInForm (#6429)
Remove the SignInForm component which is not used by studio-base
workflows. This change does away with `useInitialDeepLinkState` which
tried to do various detections for signin and moves the logic for
loading the data source and time from the url into <Workspace>.

With the changes in https://github.com/foxglove/studio/pull/6284 I don't
believe there is an issue around holding on to old "seek" functions
since those will rotate references when the player changes.

Former-commit-id: 93f3bad13e7899795863fbb5f6df4f8ba87d670a