Remove Layout ID from AppBar (#6494)
The AddPanel button on the app bar becomes disabled when there is no
active layout. This change updates the logic for detecting whether there
is an active layout to check the layout being defined rather than
tracking the _id_. If there is a defined layout that is sufficient for
the add panel button to work.

Former-commit-id: 9cfc6569ff4ce0eeb1769c6b9ba585ae1b246063