Update plots when global variables change (#6518)
Fix an issue with plots not updating when global variables are changed.

Include changes in global variables in our determination that datasets
need to be rebuilt. This is an unsophisticated implementation that
rebuilds on any global variable change. If we want to spend more time on
this we can parse the series paths and only rebuild when a variable
referenced in the paths changes.


Former-commit-id: f414b0634527ce151dc6af3da9dc3547f33baf6c