Fix unregistering event handlers across the electron context bridge (#6514)
The native app menu handlers are setup in `Workspace` using the
`useNativeAppMenuEvent` hook. This hook registers and unregisters the
handler via a `useEffect`. However, the api provided for unregistering
the handler does not work across the context bridge between renderer and
preloader. Electron uses structured cloning for the context bridge (
https://www.electronjs.org/docs/latest/api/context-bridge#parameter--error--return-type-support)
which results in copies of values on the receiving side.

This means that a caller in renderer that calls a register handler API
like `.on(event, handler);` causes the preloader (the other side of the
bridge) to receive a `handler` function that is not the same instance as
the one seen in the renderer side. So when renderer goes to call
`.off(event, handler);` and provides the same `handler` function
instance to `.off` that it provided to `.on`, the preloader sees proxies
of the function and none of the copies equal each other. The result is
that the renderer call to `.off` fails to un-register the event listener
and we end up with multiple handlers for the listener being registered.
This is what caused the many windows to open.

This PR fixes the issue by changing the `.on(event, handler)` API to
return an unregister function. The returned function from preloader
closes over the correct instance of the _handler_ to use for removing
the event handler when called. The renderer side of the bridge can call
this proxy function to unregister the correct handler.
Former-commit-id: 951aa9bb33cf41362c49c80de7bb66a9b23749f8